enhancements and fixes

+++ resources/js/screens/patients-files/show.vue
@@ -150,7 +150,7 @@
                         </a>
                         <a v-if="!payment.isEdit && !payment.isPayDebtOpened" href="#"
                            @click="deletePayment(payment.id)"
-                           class="py-1 inline-flex h-12 px-2 text-sm text-center text-red-600 transition-colors duration-200 transform lg:h-8 hover:text-red-700 focus:outline-none">
+                                     class="py-1 inline-flex h-12 px-2 text-sm text-center text-red-600 transition-colors duration-200 transform lg:h-8 hover:text-red-700 focus:outline-none">
                             <icon-delete
                                 size="5"
                                 class="transition-colors"
@@ -167,7 +167,7 @@
                         <async-button v-if="payment.isEdit || payment.isPayDebtOpened" @click="savePayment(payment)"
                                       :loading="submitted"
                                       class="ml-4 py-1 items-center justify-center h-12 px-4 text-sm text-center text-white bg-green-500 transition-colors duration-200 transform border rounded-lg lg:h-8 hover:bg-green-600 focus:outline-none">
-                            تعديل
+                            {{ payment.isPayDebtOpened ? 'حفظ' : 'تعديل' }}
                         </async-button>
                         <a v-if="payment.isEdit || payment.isPayDebtOpened" href="#" @click="cancelPayment(payment)"
                            class="ml-4 py-1 items-center justify-center h-12 px-4 text-sm text-center text-white bg-red-600 transition-colors duration-200 transform border rounded-lg lg:h-8 hover:bg-red-700 focus:outline-none">
@@ -177,6 +177,82 @@
                 </tr>
                 </tbody>
             </table>
+        </div>
+        <div
+            v-if="opened"
+            :class="`fixed z-10 inset-0 overflow-y-auto `"
+            aria-labelledby="modal-title"
+            role="dialog" aria-modal="true">
+            <div class="flex items-end justify-center min-h-screen pt-4 px-4 pb-20 text-center sm:block sm:p-0">
+                <!--
+                  Background overlay, show/hide based on modal state.
+
+                  Entering: "ease-out duration-300"
+                    From: "opacity-0"
+                    To: "opacity-100"
+                  Leaving: "ease-in duration-200"
+                    From: "opacity-100"
+                    To: "opacity-0"
+                -->
+                <div
+                    :class="`fixed inset-0 bg-gray-500 transition-opacity duration-200 ${opened?'bg-opacity-75':'bg-opacity-0'}`"
+                    @click="opened= false"
+                    aria-hidden="true"></div>
+
+                <!-- This element is to trick the browser into centering the modal contents. -->
+                <span class="hidden sm:inline-block sm:align-middle sm:h-screen" aria-hidden="true">&#8203;</span>
+
+                <!--
+                  Modal panel, show/hide based on modal state.
+
+                  Entering: "ease-out duration-300"
+                    From: "opacity-0 translate-y-4 sm:translate-y-0 sm:scale-95"
+                    To: "opacity-100 translate-y-0 sm:scale-100"
+                  Leaving: "ease-in duration-200"
+                    From: "opacity-100 translate-y-0 sm:scale-100"
+                    To: "opacity-0 translate-y-4 sm:translate-y-0 sm:scale-95"
+                -->
+                <div
+                    :class="`inline-block w-full align-bottom bg-white rounded-lg text-right overflow-hidden shadow-xl transform transition-all sm:my-8 sm:align-middle sm:max-w-xl sm:w-full duration-200  ${opened?'scale-100':'scale-0'}` ">
+                    <div class="bg-white px-4 pt-5 pb-4 sm:p-6 sm:pb-4">
+                        <div class="sm:flex sm:items-start">
+                            <div
+                                class="mx-auto flex-shrink-0 flex items-center justify-center h-12 w-12 rounded-full bg-red-100 sm:mx-0 sm:h-10 sm:w-10">
+                                <!-- Heroicon name: outline/exclamation -->
+                                <svg class="h-6 w-6 text-red-600" xmlns="http://www.w3.org/2000/svg" fill="none"
+                                     viewBox="0 0 24 24" stroke="currentColor" aria-hidden="true">
+                                    <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2"
+                                          d="M12 9v2m0 4h.01m-6.938 4h13.856c1.54 0 2.502-1.667 1.732-3L13.732 4c-.77-1.333-2.694-1.333-3.464 0L3.34 16c-.77 1.333.192 3 1.732 3z"/>
+                                </svg>
+                            </div>
+                            <div class="mt-3 text-center sm:mt-0 sm:mr-4 sm:text-right">
+                                <h3 class="text-lg leading-6 font-medium text-gray-900" id="modal-title">
+                                    حذف دفعة
+                                </h3>
+                                <div class="mt-2">
+                                    <p class="text-sm text-gray-500">
+                                        هل أنت متأكد من حذف هذه الدفعة؟
+                                    </p>
+                                </div>
+                            </div>
+                        </div>
+                    </div>
+                    <div class="bg-gray-50 px-4 py-3 sm:px-6 sm:flex sm:flex-row">
+                        <async-button
+                            type="button"
+                            :loading="submitted"
+                            @click="deletePayment(payment_id)"
+                            class="mt-3 w-full inline-flex justify-center rounded-md border border-gray-300 shadow-sm px-4 py-2 bg-red-600 text-base font-medium text-white hover:bg-red-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-red-500 sm:mr-3 sm:w-auto sm:text-sm"
+                        >
+                            حذف
+                        </async-button>
+                        <button type="button" @click="opened= false"
+                                class="mt-3 w-full inline-flex justify-center rounded-md border border-gray-300 shadow-sm px-4 py-2 bg-white text-base font-medium text-gray-700 hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500 sm:mr-3 sm:w-auto sm:text-sm">
+                            إلغاء
+                        </button>
+                    </div>
+                </div>
+            </div>
         </div>
     </div>
 </template>
@@ -199,6 +275,7 @@ export default {
             isFormManipulating: false,
             submitted: false,
             isEdit: false,
+            opened: false,
             patient_name: '',
             patient_file_number: '',
             data: [],
@@ -218,6 +295,9 @@ export default {
         this.id = this.$route.params.id;
         this.type = this.$route.query.type;
         this.getData()
+        bus.$on('item-deleted', function () {
+            this.getData();
+        })
     },
     methods: {
         back() {
@@ -292,6 +372,11 @@ export default {
             })
         },
         deletePayment(id) {
+            this.payment_id = id;
+            if (!this.opened){
+                this.opened = true;
+                return;
+            }
             axios.delete(`/api/patients-files/${id}`).then(({data}) => {
                 bus.$emit('flash-message', {text: data.message, type: 'success'});
                 this.getData();
@@ -299,6 +384,7 @@ export default {
                 bus.$emit('flash-message', {text: response.data.message, type: 'error'});
             }).finally(() => {
                 this.submitted = false;
+                this.opened = false;
             })
         },
         editPayment(payment) {
